src/agent/TradeFlowAgent/requirements.txt
# LLM 支持
litellm>=1.0.0

# 环境变量管理
python-dotenv>=1.0.0

# 异步编程支持 (asyncio 是 Python 标准库，不需要安装)

# HTTP 客户端 (用于 API 调用)
httpx>=0.25.0
aiohttp>=3.9.0

# JSON 处理
orjson>=3.9.0

# 日志和调试
loguru>=0.7.0

# 测试框架
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.12.0

# 开发工具
black>=23.0.0
flake8>=6.1.0
mypy>=1.5.0

# 类型提示
typing-extensions>=4.8.0

# Web 框架（用于提供 HTTP API）
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0

# Google ADK
google-adk>=0.1.0

# 数据分析和可视化（用于代码执行）
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0